requirements.txt
python-dotenv
faiss-cpu
pymupdf
langchain-community
langchain-openai
langchain-core
streamlit